rust: rename agama-dbus-server:{server,web}
