+++ rust/agama-dbus-server/src/lib.rs
@@ -2,5 +2,5 @@ pub mod error;
 pub mod l10n;
 pub mod network;
 pub mod questions;
-pub mod server;
-pub use server::service;
+pub mod web;
+pub use web::service;
